RavenDB-15821 - Cannot Test Credentials when setting up Server-Wide backup
