+++ src/Raven.Server/Web/System/OngoingTasksHandler.cs
@@ -289,84 +289,6 @@ namespace Raven.Server.Web.System
             return taskInfo;
         }
 
-        [RavenAction("/databases/*/admin/periodic-backup/test-credentials", "POST", AuthorizationStatus.DatabaseAdmin)]
-        public async Task TestPeriodicBackupCredentials()
-        {
-            var type = GetQueryStringValueAndAssertIfSingleAndNotEmpty("type");
-
-            if (Enum.TryParse(type, out PeriodicBackupConnectionType connectionType) == false)
-                throw new ArgumentException($"Unknown backup connection: {type}");
-
-            using (ServerStore.ContextPool.AllocateOperationContext(out TransactionOperationContext context))
-            {
-                DynamicJsonValue result;
-                try
-                {
-                    var connectionInfo = await context.ReadForMemoryAsync(RequestBodyStream(), "test-connection");
-                    switch (connectionType)
-                    {
-                        case PeriodicBackupConnectionType.S3:
-                            var s3Settings = JsonDeserializationClient.S3Settings(connectionInfo);
-                            using (var awsClient = new RavenAwsS3Client(s3Settings, logger: Logger, cancellationToken: ServerStore.ServerShutdown))
-                            {
-                                awsClient.TestConnection();
-                            }
-                            break;
-                        case PeriodicBackupConnectionType.Glacier:
-                            var glacierSettings = JsonDeserializationClient.GlacierSettings(connectionInfo);
-                            using (var glacierClient = new RavenAwsGlacierClient(glacierSettings, logger: Logger, cancellationToken: ServerStore.ServerShutdown))
-                            {
-                                glacierClient.TestConnection();
-                            }
-                            break;
-                        case PeriodicBackupConnectionType.Azure:
-                            var azureSettings = JsonDeserializationClient.AzureSettings(connectionInfo);
-                            using (var azureClient = new RavenAzureClient(azureSettings, logger: Logger, cancellationToken: ServerStore.ServerShutdown))
-                            {
-                                azureClient.TestConnection();
-                            }
-                            break;
-                        case PeriodicBackupConnectionType.GoogleCloud:
-                            var googleCloudSettings = JsonDeserializationClient.GoogleCloudSettings(connectionInfo);
-                            using (var googleCloudClient = new RavenGoogleCloudClient(googleCloudSettings, cancellationToken: ServerStore.ServerShutdown))
-                            {
-                                await googleCloudClient.TestConnection();
-                            }
-                            break;
-                        case PeriodicBackupConnectionType.FTP:
-                            var ftpSettings = JsonDeserializationClient.FtpSettings(connectionInfo);
-                            using (var ftpClient = new RavenFtpClient(ftpSettings))
-                            {
-                                ftpClient.TestConnection();
-                            }
-                            break;
-                        case PeriodicBackupConnectionType.Local:
-                        case PeriodicBackupConnectionType.None:
-                        default:
-                            throw new ArgumentOutOfRangeException();
-                    }
-
-                    result = new DynamicJsonValue
-                    {
-                        [nameof(NodeConnectionTestResult.Success)] = true,
-                    };
-                }
-                catch (Exception e)
-                {
-                    result = new DynamicJsonValue
-                    {
-                        [nameof(NodeConnectionTestResult.Success)] = false,
-                        [nameof(NodeConnectionTestResult.Error)] = e.ToString()
-                    };
-                }
-
-                using (var writer = new BlittableJsonTextWriter(context, ResponseBodyStream()))
-                {
-                    context.Write(writer, result);
-                }
-            }
-        }
-
         [RavenAction("/databases/*/admin/periodic-backup/config", "GET", AuthorizationStatus.DatabaseAdmin)]
         public Task GetConfiguration()
         {
