+++ src/Raven.Studio/typescript/viewmodels/manage/editServerWideBackup.ts
@@ -187,7 +187,7 @@ class editServerWideBackup extends viewModelBase {
         bs.isTestingCredentials(true);
         bs.testConnectionResult(null);
 
-        new testPeriodicBackupCredentialsCommand(this.activeDatabase(), bs.connectionType, bs.toDto())
+        new testPeriodicBackupCredentialsCommand(bs.connectionType, bs.toDto())
             .execute()
             .done((result: Raven.Server.Web.System.NodeConnectionTestResult) => {
                 bs.testConnectionResult(result);
